refactor: move validate_evaluation_results to test_utils
