+++ docling_eval/utils/utils.py
@@ -36,7 +36,6 @@ from docling_eval.datamodels.types import (
     EvaluationModality,
     PredictionProviderType,
 )
-from docling_eval.evaluators.ocr_evaluator import OcrDatasetEvaluationResult
 
 
 def get_binhash(binary_data: bytes) -> str:
@@ -633,38 +632,6 @@ def modalities_of_prediction_type(
         return None
 
     return prediction_provider_class.prediction_modalities
-
-
-def validate_evaluation_results(
-    target_path: Path,
-    benchmark: str,
-    modality: str,
-    evaluation_type: str = "ocr",
-) -> OcrDatasetEvaluationResult:
-    eval_json_filename = f"evaluation_{benchmark}_{modality}.json"
-    eval_json_path = target_path / "evaluations" / evaluation_type / eval_json_filename
-
-    assert eval_json_path.exists(), f"Evaluation JSON file not found: {eval_json_path}"
-
-    with open(eval_json_path, "r") as f:
-        result = json.load(f)
-
-    assert result is not None, "Evaluation JSON file is empty or invalid."
-    assert result, "Overall metrics not found in evaluation results."
-
-    metrics = OcrDatasetEvaluationResult(**result)
-
-    assert (
-        metrics.f1_score > 0
-    ), f"F1 score ({metrics.f1_score}) must be greater than 0."
-    assert (
-        metrics.precision > 0
-    ), f"Precision score ({metrics.precision}) must be greater than 0."
-    assert (
-        metrics.recall > 0
-    ), f"Recall score ({metrics.recall}) must be greater than 0."
-
-    return metrics
 
 
 def does_intersection_area_exceed_threshold(
